// controlplane/src/core/clickhouse/client/enums/ClickHouseDataFormat.ts
export enum ClickHouseDataFormat {
  TabSeparated = 'TabSeparated',
  TabSeparatedRaw = 'TabSeparatedRaw',
  TabSeparatedWithNames = 'TabSeparatedWithNames',
  TabSeparatedWithNamesAndTypes = 'TabSeparatedWithNamesAndTypes',
  TabSeparatedRawWithNames = 'TabSeparatedRawWithNames',
  TabSeparatedRawWithNamesAndTypes = 'TabSeparatedRawWithNamesAndTypes',
  CSV = 'CSV',
  CSVWithNames = 'CSVWithNames',
  CSVWithNamesAndTypes = 'CSVWithNamesAndTypes',
  CustomSeparated = 'CustomSeparated',
  CustomSeparatedWithNames = 'CustomSeparatedWithNames',
  CustomSeparatedWithNamesAndTypes = 'CustomSeparatedWithNamesAndTypes',
  Values = 'Values',
  Vertical = 'Vertical',
  TSV = 'TSV',
  JSON = 'JSON',
  JSONStrings = 'JSONStrings',
  JSONCompact = 'JSONCompact',
  JSONCompactStrings = 'JSONCompactStrings',
  TSKV = 'TSKV',
  Pretty = 'Pretty',
  PrettyNoEscapes = 'PrettyNoEscapes',
  PrettyCompact = 'PrettyCompact',
  PrettyCompactNoEscapes = 'PrettyCompactNoEscapes',
  PrettyCompactMonoBlock = 'PrettyCompactMonoBlock',
  PrettySpace = 'PrettySpace',
  PrettySpaceNoEscapes = 'PrettySpaceNoEscapes',
  Null = 'Null',
  LineAsString = 'LineAsString',
}